Add Package Parameter
It is now possible to initialize a standard package using
`velocitas init -p devenv-runtimes`. This command will depending on if the
devenv-runtimes package is defined in .velocitas.json use the version specified
in .velocitas.json or use the latest version if not specified in .velocitas.json.

Using a custom version:
`velocitas init -p devenv-runtimes@v3.0.0`.

Using a custom package:
`velocitas init -p https://github.com/SoftwareDefinedVehicle/devenv-runtimes.git@v3.0.0`

If initializing a package which is not part of .velocitas.json it will
automatically be added to it. If a package specifies components but no
component is referenced in .velocitas.json, all components will be added
automatically during initialization.